Teste find

+++ src/com/devhealth/dao/PessoaFisicaDAO.java
@@ -2,7 +2,7 @@ package com.devhealth.dao;
 
 import com.devhealth.entity.PessoaFisica;
 
-public class PessoaFisicaDAO extends GenericDAO<PessoaFisica, Long> {
+public class PessoaFisicaDAO extends GenericDAO<PessoaFisica, Integer> {
 
 	public PessoaFisicaDAO() {
 		super(PessoaFisica.class);
