+++ src/com/devhealth/dao/ProcedimentoDAO.java
@@ -2,7 +2,7 @@ package com.devhealth.dao;
 
 import com.devhealth.entity.Procedimento;
 
-public class ProcedimentoDAO extends GenericDAO<Procedimento, Long> {
+public class ProcedimentoDAO extends GenericDAO<Procedimento, Integer> {
 	public ProcedimentoDAO() {
 		super(Procedimento.class);
 	}
